refactored so that pages pull components from the components folder to make code more compartmentalized and easier to change
