+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react"
-import Access from "./components/auth/Access"
-import Feed from "./components/feed/feed" // Ensure the correct path for Feed
+import Access from "./pages/Access"
+import Feed from "./pages/feed" // Ensure the correct path for Feed
 
 const App: React.FC = () => {
   return <Feed />
